Add LivestreamChannelLogic.kt with bidirectional pagination and no DB.
