+++ stream-chat-android-state/src/main/java/io/getstream/chat/android/state/plugin/listener/internal/QueryChannelListenerState.kt
@@ -73,9 +73,7 @@ internal class QueryChannelListenerState(private val logic: LogicRegistry) : Que
             "[onQueryChannelResult] cid: $channelType:$channelId, " +
                 "request: $request, result: ${result.stringify { it.cid }}"
         }
-        val channelStateLogic = logic.channel(channelType, channelId).stateLogic()
-
-        result.onSuccess { channel -> channelStateLogic.propagateChannelQuery(channel, request) }
-            .onError(channelStateLogic::propagateQueryError)
+        val channelLogic = logic.channel(channelType, channelId)
+        channelLogic.onQueryChannelResult(request, result)
     }
 }
