+++ stream-chat-android-state/src/main/java/io/getstream/chat/android/state/plugin/state/StateRegistry.kt
@@ -26,7 +26,6 @@ import io.getstream.chat.android.models.Message
 import io.getstream.chat.android.models.User
 import io.getstream.chat.android.models.querysort.QuerySorter
 import io.getstream.chat.android.state.event.handler.internal.batch.BatchEvent
-import io.getstream.chat.android.state.plugin.config.MessageLimitConfig
 import io.getstream.chat.android.state.plugin.state.channel.internal.ChannelMutableState
 import io.getstream.chat.android.state.plugin.state.channel.thread.ThreadState
 import io.getstream.chat.android.state.plugin.state.channel.thread.internal.ThreadMutableState
@@ -58,7 +57,6 @@ public class StateRegistry constructor(
     private val job: Job,
     private val now: () -> Long,
     private val scope: CoroutineScope,
-    private val messageLimitConfig: MessageLimitConfig,
 ) {
 
     private val logger by taggedLogger("Chat:StateRegistry")
@@ -113,7 +111,6 @@ public class StateRegistry constructor(
                 userStateFlow,
                 latestUsers,
                 activeLiveLocations,
-                getMessageLimitFilter(channelType),
                 now,
             )
         }
@@ -202,17 +199,5 @@ public class StateRegistry constructor(
             it.destroy()
         }
         logger.i { "[removeChanel] removed channel($channelType, $channelId): $removed" }
-    }
-
-    private fun getMessageLimitFilter(channelType: String): (Collection<Message>) -> Collection<Message> =
-        messageLimitConfig.channelMessageLimits.firstOrNull { it.channelType == channelType }
-            ?.let { createMessageLimitFilter(it.limit) }
-            ?: noOpMessagesLimitFilter
-
-    private fun createMessageLimitFilter(limit: Int): (Collection<Message>) -> Collection<Message> = {
-        when (it.size > limit) {
-            true -> it.sortedBy { it.createdAt ?: it.createdLocallyAt }.takeLast(limit)
-            false -> it
-        }
     }
 }
