+++ stream-chat-android-state/src/test/java/io/getstream/chat/android/state/plugin/state/channel/internal/ChannelMutableStateTests.kt
@@ -52,7 +52,6 @@ internal class ChannelMutableStateTests {
             activeLiveLocations = MutableStateFlow(
                 emptyList(),
             ),
-            messagesLimitFilter = { it },
             now = ChannelMutableStateTests::currentTime,
         )
     }
